requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# LlamaIndex core
llama-index-core>=0.12.0

# LLM providers
llama-index-llms-google-genai>=0.8.0
llama-index-llms-openai>=0.3.0

# Embeddings
llama-index-embeddings-huggingface>=0.5.0

# Vector store — Chroma (works with llama-index-core 0.14+ and Python 3.14).
# Qdrant integration wheels for 3.14 conflict with core; use Python 3.11–3.13 + llama-index-vector-stores-qdrant if you need Qdrant.
llama-index-vector-stores-chroma>=0.4.0
chromadb>=0.5.0

# Neo4j driver (KG import: scripts/load_kg_json_to_neo4j.py)
neo4j>=5.0.0

# PDF processing
pymupdf>=1.24.0
pdfplumber>=0.11.0

# Retrieval & reranking
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# Utilities
torch>=2.2.0
aiofiles>=24.1.0
